+++ src/app/shared/pull-data/pull-data.service.spec.ts
@@ -21,42 +21,129 @@ describe('PullDataService', () => {
   describe('combineObservationCodes', () => {
     it('should change duplicate display names', () => {
       expect(
-        service.combineObservationCodes([
-          {
-            coding: [
-              {
-                code: 'phv00493202.v1.p1',
-                system: ''
-              }
-            ],
-            datatype: 'CodeableConcept',
-            items: ['Affection status']
-          },
-          {
-            coding: [
-              {
-                code: 'phv00492021.v1.p1',
-                system: 'some-system'
-              }
-            ],
-            datatype: 'CodeableConcept',
-            items: ['Affection status']
-          }
-        ])
+        service.getCodesFromCriteria({
+          condition: 'and',
+          rules: [
+            {
+              condition: 'and',
+              rules: [
+                {
+                  field: {
+                    element: 'code text',
+                    value: '',
+                    selectedObservationCodes: {
+                      coding: [
+                        {
+                          code: '3137-7',
+                          system: 'http://loinc.org'
+                        },
+                        {
+                          code: '11881-0',
+                          system: 'http://loinc.org'
+                        }
+                      ],
+                      datatype: 'Quantity',
+                      defaultUnit: 'cm',
+                      defaultUnitSystem: 'http://unitsofmeasure.org',
+                      items: ['Height cm', 'Fundus Height']
+                    }
+                  }
+                }
+              ],
+              resourceType: 'Observation'
+            },
+            {
+              condition: 'and',
+              rules: [
+                {
+                  field: {
+                    element: 'code text',
+                    value: '',
+                    selectedObservationCodes: {
+                      coding: [
+                        {
+                          code: '3137-7',
+                          system: 'http://loinc.org'
+                        },
+                        {
+                          code: '8302-2',
+                          system: 'http://loinc.org'
+                        }
+                      ],
+                      datatype: 'Quantity',
+                      defaultUnit: 'cm',
+                      defaultUnitSystem: 'http://unitsofmeasure.org',
+                      items: ['Height cm', 'Height Peds']
+                    }
+                  }
+                }
+              ],
+              resourceType: 'Observation'
+            },
+            {
+              condition: 'and',
+              rules: [
+                {
+                  field: {
+                    element: 'code',
+                    value: {
+                      codes: ['27250050000320', '27250050000320', '27250050000340'],
+                      items: ['METFORMIN     TAB 500MG', 'METFORMIN    TAB 500MG', 'METFORMIN   TAB 850MG']
+                    }
+                  }
+                }
+              ],
+              resourceType: 'MedicationDispense'
+            },
+            {
+              condition: 'and',
+              rules: [
+                {
+                  field: {
+                    element: 'code',
+                    value: {
+                      codes: ['27250050000320', '27250050000340', '27250050000340'],
+                      items: ['METFORMIN     TAB 500MG', 'METFORMIN    TAB 850MG', 'METFORMIN   TAB 850MG']
+                    }
+                  }
+                }
+              ],
+              resourceType: 'MedicationDispense'
+            }
+          ]
+        })
       ).toEqual({
-        coding: [
-          {
-            code: 'phv00493202.v1.p1',
-            system: ''
-          },
-          {
-            code: 'phv00492021.v1.p1',
-            system: 'some-system'
-          }
-        ],
-        // "datatype" is not used in the pull data step
-        datatype: 'any',
-        items: ['Affection status | phv00493202.v1.p1', 'Affection status | phv00492021.v1.p1 | some-system']
+        Observation: {
+          coding: [
+            {
+              code: '3137-7',
+              system: 'http://loinc.org'
+            },
+            {
+              code: '11881-0',
+              system: 'http://loinc.org'
+            },
+            {
+              code: '8302-2',
+              system: 'http://loinc.org'
+            }
+          ],
+          items: ['Height cm', 'Fundus Height', 'Height Peds'],
+          datatype: 'any'
+        },
+        MedicationDispense: {
+          codes: ['27250050000320', '27250050000320', '27250050000340', '27250050000340'],
+          items: [
+            'METFORMIN     TAB 500MG',
+            'METFORMIN    TAB 500MG',
+            'METFORMIN   TAB 850MG',
+            'METFORMIN    TAB 850MG'
+          ]
+        },
+        MedicationRequest: {
+          codes: [],
+          items: []
+        }
       });
     });
   });